pyproject.toml
[tool.poetry]
name = "testwhynot"
version = "0.1.0"
description = ""
authors = ["Artyom <kostyuchenko.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
